+++ packages/cli-plugin-build/src/index.ts
@@ -6,6 +6,7 @@ import { copyFiles } from '@midwayjs/faas-code-analysis';
 import * as ts from 'typescript';
 export class BuildPlugin extends BasePlugin {
   isMidwayHooks = false;
+  private midwayBinBuild: { include?: string[] } = {};
   commands = {
     build: {
       lifecycleEvents: [
@@ -14,6 +15,7 @@ export class BuildPlugin extends BasePlugin {
         'copyFile',
         'compile',
         'emit',
+        'complete',
       ],
       options: {
         clean: {
@@ -31,10 +33,16 @@ export class BuildPlugin extends BasePlugin {
           usage: 'build out path',
         },
         tsConfig: {
-          usage: 'tsConfig json file path',
+          usage: 'json string / file path / object',
         },
         buildCache: {
           usage: 'save build cache',
+        },
+        exclude: {
+          usage: 'copy file exclude',
+        },
+        include: {
+          usage: 'copy file include',
         },
       },
     },
@@ -46,6 +54,7 @@ export class BuildPlugin extends BasePlugin {
     'build:copyFile': this.copyFile.bind(this),
     'build:compile': this.compile.bind(this),
     'build:emit': this.emit.bind(this),
+    'build:complete': this.complete.bind(this),
   };
 
   private compilerHost: CompilerHost;
@@ -63,6 +72,12 @@ export class BuildPlugin extends BasePlugin {
       if (config.source) {
         this.options.srcDir = config.source;
       }
+    }
+
+    const packageJsonPath = join(this.core.cwd, 'package.json');
+    if (existsSync(packageJsonPath)) {
+      const pkgJson = JSON.parse(readFileSync(packageJsonPath).toString());
+      this.midwayBinBuild = pkgJson['midway-bin-build'] || {};
     }
   }
 
@@ -91,17 +106,46 @@ export class BuildPlugin extends BasePlugin {
   }
 
   async copyFile() {
-    const targetDir = this.getOutDir();
-    this.core.debug('CopyFile TargetDir', targetDir);
+    const outDir = this.getOutDir();
+    this.core.debug('CopyFile TargetDir', outDir);
+    const exclude = this.options.exclude ? this.options.exclude.split(',') : [];
+    const sourceDir = join(this.core.cwd, this.options.srcDir || 'src');
+    const targetDir = join(this.core.cwd, outDir);
     await copyFiles({
-      sourceDir: join(this.core.cwd, this.options.srcDir || 'src'),
-      targetDir: join(this.core.cwd, targetDir),
-      defaultInclude: ['**/*'],
-      exclude: ['**/*.ts', '**/*.js'],
+      sourceDir,
+      targetDir,
+      defaultInclude: this.midwayBinBuild.include
+        ? this.midwayBinBuild.include
+        : ['**/*'],
+      exclude: ['**/*.ts', '**/*.js'].concat(exclude),
       log: path => {
         this.core.cli.log(`   - Copy ${path}`);
       },
     });
+
+    // midway core DEFAULT_IGNORE_PATTERN
+    let include = [
+      '**/public/**/*.js',
+      '**/view/**/*.js',
+      '**/views/**/*.js',
+      '**/app/extend/**/*.js',
+    ];
+    if (this.options.include !== undefined) {
+      include = this.options.include ? this.options.include.split(',') : [];
+    }
+
+    if (include.length) {
+      await copyFiles({
+        sourceDir,
+        targetDir,
+        defaultInclude: include,
+        exclude,
+        log: path => {
+          this.core.cli.log(`   - Copy ${path}`);
+        },
+      });
+    }
+    this.core.cli.log('   - Copy Complete');
   }
 
   async compile() {
@@ -195,9 +239,13 @@ export class BuildPlugin extends BasePlugin {
   private getTsConfig() {
     const { cwd } = this.core;
     this.core.debug('CWD', cwd);
-    const { tsConfig } = this.options;
+    let { tsConfig } = this.options;
     let tsConfigResult;
     if (typeof tsConfig === 'string') {
+      // if ts config is file
+      if (existsSync(tsConfig)) {
+        tsConfig = readFileSync(tsConfig).toString();
+      }
       try {
         tsConfigResult = JSON.parse(tsConfig);
       } catch (e) {
@@ -221,5 +269,9 @@ export class BuildPlugin extends BasePlugin {
       }
     }
     return tsConfigResult;
+  }
+
+  async complete() {
+    this.core.cli.log('Build Complete!');
   }
 }