+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.74](https://github.com/midwayjs/midway/compare/v1.2.73...v1.2.74) (2021-07-06)
+
+**Note:** Version bump only for package @midwayjs/cli
+
+
+
+
+
+## [1.2.73](https://github.com/midwayjs/midway/compare/v1.2.73-beta.1...v1.2.73) (2021-06-25)
+
+**Note:** Version bump only for package @midwayjs/cli
+
+
+
+
+
+## [1.2.72](https://github.com/midwayjs/midway/compare/v1.2.71...v1.2.72) (2021-06-18)
+
+
+### Bug Fixes
+
+* deploy with env ([#128](https://github.com/midwayjs/midway/issues/128)) ([bcfbff8](https://github.com/midwayjs/midway/commit/bcfbff8e70f9707695f2a837f4c6f7c55cb56d38))
+
+
+
+
+
+## [1.2.71](https://github.com/midwayjs/midway/compare/v1.2.70...v1.2.71) (2021-06-07)
+
+**Note:** Version bump only for package @midwayjs/cli
+
+
+
+
+
+## [1.2.70](https://github.com/midwayjs/midway/compare/v1.2.69...v1.2.70) (2021-06-02)
+
+**Note:** Version bump only for package @midwayjs/cli
+
+
+
+
+
+## [1.2.69](https://github.com/midwayjs/midway/compare/v1.2.68...v1.2.69) (2021-06-01)
+
+**Note:** Version bump only for package @midwayjs/cli
+
+
+
+
+
 ## [1.2.68](https://github.com/midwayjs/midway/compare/v1.2.67...v1.2.68) (2021-05-13)